fix: manage cast on iOS side (major and minor in BeaconData)

+++ packages/drivekit_trip_analysis/flutter_drivekit_trip_analysis_ios/ios/flutter_drivekit_trip_analysis_ios/Sources/flutter_drivekit_trip_analysis_ios/PigeonMapper.swift
@@ -291,8 +291,8 @@ extension PigeonBeaconData {
     init(from beacon: BeaconData) {
         self.init(
             proximityUuid: beacon.proximityUuid,
-            major: beacon.major,
-            minor: beacon.minor
+            major: Int64(beacon.major),
+            minor: Int64(beacon.minor)
         )
     }
 }
